Move Bootstrap to presto-main
